Split loadgen operation into shared file
